05-data-platforms/my-dlt-pipeline/requirements.txt
dlt[duckdb]>=1.22.1
marimo>=0.8.0
ibis-framework>=9.0.0
plotly>=5.0.0